+++ src/nio/responses.py
@@ -1381,6 +1381,7 @@ class RoomEventRelationsResponse(Response):
     events: list[Event]
     prev_batch: str | None
     next_batch: str | None
+    recursion_depth: int | None = None
 
     @classmethod
     @verify(
@@ -1392,7 +1393,8 @@ class RoomEventRelationsResponse(Response):
         events = [Event.parse_event(e) for e in parsed_dict["chunk"]]
         prev_batch = parsed_dict.get("prev_batch")
         next_batch = parsed_dict.get("next_batch")
-        return cls(room_id, event_id, events, prev_batch, next_batch)
+        recursion_depth = parsed_dict.get("recursion_depth")
+        return cls(room_id, event_id, events, prev_batch, next_batch, recursion_depth)
 
 
 @dataclass